Move to modules instead of classes for home page and common interface
