+++ data_explorer/app/config.py
@@ -1,12 +1,9 @@
 """General configuration for the app."""
 
 SESSION_STATE_VARIABLES = [
-    "base_path",
-    "pipeline",
     "selected_pipeline_path",
     "run",
     "selected_run_path",
-    "available_runs",
     "component",
     "selected_component_path",
     "partition",
