+++ data_explorer/app/pages/images.py
@@ -3,7 +3,9 @@ import io
 
 import streamlit as st
 from df_helpers.fields import get_image_fields
+from interfaces.common import render_common_interface
 from interfaces.dataset_interface import DatasetLoaderApp
+from interfaces.sidebar import render_sidebar
 from PIL import Image
 
 
@@ -26,8 +28,10 @@ class ImageGalleryApp(DatasetLoaderApp):
                 cols[i % 5].image(image, use_column_width=True)
 
 
+render_sidebar()
+render_common_interface()
+
 app = ImageGalleryApp()
-app.create_common_interface()
 field_mapping, selected_fields = app.get_fields_mapping()
 dask_df = app.load_dask_dataframe(field_mapping)
 df = app.load_pandas_dataframe(dask_df, field_mapping)
